Small tweaks, specifically check that fft_size is 2^n between 32 and 8192 instead of a multiple of 32, and make matlab exported data keep full precision on frequencies instead of clipping them to ints.

+++ USRPAnalyzer/usrp_analyzer.py
@@ -368,7 +368,7 @@ class top_block(gr.top_block):
         #   ...
         # ]
         matlab_format_data = np.zeros(self.cfg.nsteps, dtype=[
-            ('center_frequency', np.uint32),
+            ('center_frequency', np.float64),
             ('samples', np.complex64, (chunk_size,))
         ])
 
@@ -403,7 +403,7 @@ class top_block(gr.top_block):
         #   ...
         # ]
         matlab_format_data = np.zeros(len(self.cfg.bin_freqs), dtype=[
-            ('bin_frequency', np.uint32),
+            ('bin_frequency', np.float64),
             ('samples', np.complex64, (self.cfg.n_averages,))
         ])
